+++ create-echoed/src/ansi.ts
@@ -1,0 +1,3 @@
+export const AnsiReset = "\x1b[0m";
+export const AnsiRed = "\x1b[31m";
+export const AnsiGreen = "\x1b[32m";