+++ create-echoed/src/creator.ts
@@ -3,12 +3,12 @@ import path from "path";
 
 export class Creator {
   constructor(
-    private echoedRootDir: string,
+    private createEchoedRootDir: string,
     private cwd: string,
   ) {}
 
-  async run(): Promise<void> {
-    const dir = path.join(this.echoedRootDir, "create", "template");
+  async createJestExample(): Promise<void> {
+    const dir = path.join(this.createEchoedRootDir, "template", "jest");
     await fs.promises.cp(dir, this.cwd, { recursive: true });
   }
 }