+++ jest.config.js
@@ -40,7 +40,6 @@ module.exports = {
   preset: "ts-jest",
   verbose: true,
   rootDir: "src",
-  testPathIgnorePatterns: ["<rootDir>/create/template"],
   moduleNameMapper: {
     "^@/(.*)$": "<rootDir>/$1",
   },