+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+### Added
+- Create "create-echoed" package to reduce size of "echoed" package
+
 ## 0.1.0
 ### Added
 - Feature to create tests from YAML